вынес data в отдельный заголовочный файл
