+++ components/components.pri
@@ -10,4 +10,5 @@ HEADERS += \
            $$PWD/src/reduce.h \
            $$PWD/src/calc.h \
            $$PWD/src/history.h \
-           $$PWD/src/convert.h
+           $$PWD/src/convert.h \
+           $$PWD/src/data.h
