+++ components/src/component.h
@@ -8,6 +8,8 @@
 #include <sstream>
 #include <boost/shared_ptr.hpp>
 
+#include "data.h"
+
 // node, router forward declaration
 namespace alpha {
 namespace protort {
@@ -38,24 +40,6 @@ struct output
 /*!
  * \brief data - структура для передачи сообщений
  */
-struct data
-{
-    float val; //значение, которое генерирует генератор
-    std::time_t time; //метка времени
-
-    //!\brief распаковывает строку и инициализирует this
-    void unpack(std::string const & str){
-        const data *d = reinterpret_cast<const data*> (str.data());
-        val = d->val;
-        time = d->time;
-    }
-
-    //!\brief запаковывает данные в строку
-    std::string pack(){
-        std::string s(reinterpret_cast<char*>(this),sizeof(data));
-        return s;
-    }
-};
 
 using output_list = std::vector<output>;
